Refactor commands into their own methods
