+++ joern-cli/frontends/rubysrc2cpg/src/main/antlr4/io/joern/rubysrc2cpg/parser/RubyParser.g4
@@ -159,10 +159,10 @@ invocationWithoutParentheses
     ;
 
 command
-    :   SUPER argumentsWithoutParentheses
-    |   YIELD argumentsWithoutParentheses
-    |   methodIdentifier argumentsWithoutParentheses
-    |   primary WS* (DOT | COLON2) wsOrNl* methodName argumentsWithoutParentheses
+    :   SUPER argumentsWithoutParentheses                                                                                       # superCommand
+    |   YIELD argumentsWithoutParentheses                                                                                       # yieldCommand
+    |   methodIdentifier argumentsWithoutParentheses                                                                            # simpleMethodCommand
+    |   primary WS* (DOT | COLON2) wsOrNl* methodName argumentsWithoutParentheses                                               # memberAccessCommand
     ;
 
 chainedCommandWithDoBlock
